// libs/shared/modules/src/notes/models/domain/note.ts
/**
 * Anecdotes API
 * Anecdotes API's
 *
 * OpenAPI spec version: 0.0.3
 * Contact: apiteam@anecdotes.ai
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { ResourceType } from 'core/modules/data/models';

export interface Note {
  /**
   * Resource id that supports notes
   */
  resource_id?: string;

  /**
   * Resource type that supports notes
   */
  resource_type?: ResourceType;

  /**
   * Note_id
   */
  note_id?: string;

  /**
   * The user who last edited the note
   */
  note_updated_by?: string;

  /**
   * The note data
   */
  note_text?: string;

  /**
   * The time the note last edited
   */
  note_timestamp?: Date;

  is_deleted?: boolean;
}